minSubArraylen function
Write a function called minSubArraylen which accepts two parameters - an array of positive integers and a positive integer
This function should return the minimal length of a contiguous subarray of which the sum is greater than or equal to the integer passed to the function. If there isnot one, return 0 instead.